reafactor(cli): rename _cli to cli
