+++ src/mecab_text_cleaner/__main__.py
@@ -1,4 +1,4 @@
-from ._cli import main
+from .cli import main
 
 if __name__ == "__main__":
     main()
